Add basic class wrapping support and some pointer stuff
The current class support is fairly limited (e.g. no move or copy
constructors yet, no inheritance, etc.), but it is a good base to
build further capabilities on top of. The pointer support (added
to allow passing class objects between C++ and D, which isn't
working yet) is similarly basic but good enough to demonstrate
simple proof-of-concept code.